Create bbitperp
This code is meant to serve as a guide and a starting point for you to build upon. It is not intended to be runnable as-is, as it contains placeholder values that you need to replace with your own bot token, chat ID, and API endpoint URLs.

To use this script, you will need to:

Obtain a bot token from Telegram by following the instructions at https://core.telegram.org/bots#6-botfather.

Find the chat ID of the recipient of the messages by following the instructions at https://stackoverflow.com/questions/32423837/telegram-bot-how-to-get-a-group-chat-id.

Replace the placeholder value of YOUR_BOT_TOKEN with your actual bot token.

Replace the placeholder value of CHAT_ID with the chat ID of the recipient.

Replace the placeholder value of BYBIT_API_URL with the Bybit API endpoint for getting the MACD indicator.

Replace the placeholder value of TELEGRAM_API_URL with the Telegram API endpoint for sending messages.

Install any required libraries, such as requests, by running pip install requests.

Run the script using a command like python script.py.

I hope this helps! Let me know if you have any questions.