Create byperpv2
The first thing we do is import the necessary libraries: requests for making HTTP requests to the API, time for sleeping the program for a minute between each iteration, and telegram for interacting with the Telegram API.

Next, we create a telegram.Bot object with our API token and set the chat_id variable to the ID of the chat where we want to send the message. We also set the timeframe variable to the number of minutes we want to use for the MACD calculation and the url variable to the URL for the BTC/USDT perpetual on Bybit.

In the while loop, we make a request to the API using the requests.get function and pass it the url and params variables. The params variable contains the symbol (BTCUSDT), interval (15 minutes), and limit (1) for the request. We set the response variable to the JSON response from the API and extract the first element of the result list, which contains the MACD and signal line values for the specified timeframe.

We then extract the macd and signal values from the data and use an if statement to check if the MACD line has crossed the signal line. If the MACD line is greater than the signal line, we set the direction variable to "upwards". If the MACD line is less than the signal line, we set the direction variable to "downwards". If the MACD line and signal line are equal, we set the direction variable to None.

Next, we use an if statement to check if the price has risen or dropped since the last message. We do this by comparing the last_price variable (which is set to the current close price) to the previous_price variable (which is initially set to None). If the last_price is greater than the previous_price, we set the price_change variable to "risen". If the last_price is less than the previous_price, we set the price_change variable to "dropped". If the last_price is equal to the previous_price, we set the price_change variable to "not changed".

Finally, we use an if statement to check if the direction variable is not None. If it is not None, this means that the MACD line has crossed the signal line and we send a message using the telegram.Bot.send_message function. The message includes the direction in which the MACD line crossed the signal line and the change in price since the last message.

After sending the message (if applicable), we use the time.sleep function to pause the program for one minute before repeating the process. This allows us to check for crossings every minute in the 15-minute timeframe.