Update byperpv2
The first thing we do is import the necessary libraries: requests for making HTTP requests to the API, telegram for interacting with the Telegram API, talib for using the crossover function to detect crossings more reliably, and asyncio for sleeping the program more efficiently.

Next, we create a telegram.Bot object with our API token and set the chat_id variable to the ID of the chat where we want to send the message. We also set the timeframe variable to the number of minutes we want to use for the MACD calculation and the url variable to the URL for the BTC/USDT perpetual on Bybit.

We define an async function called main that will contain the main logic of the program. Inside the function, we have a while loop that will run indefinitely.

Inside the loop, we make a request to the API using the requests.get function and pass it the url and params variables. The params variable contains the symbol (BTCUSDT), interval (15 minutes), and limit (1) for the request. We set the response variable to the JSON response from the API and extract the first element of the result list, which contains the MACD and signal line values for the specified timeframe.

We then extract the macd and signal values from the data and use the crossover function from the talib library to check if the MACD line has crossed the signal line. The crossover function returns an array of 1s, -1s, and 0s, where 1 indicates that the MACD line has crossed the signal line upwards, -1 indicates that it has crossed downwards, and 0 indicates that there was no crossing. We use the -1 index of the array to get the most recent crossing event and check if it is 1, -1, or 0.

Next, we use an if statement to check if the price has risen or dropped since the last message. We do this by comparing the last_price variable (which is set to the current close price) to the previous_price variable (which is initially set to None). If the last_price is greater than the previous_price, we set the price_change variable to "risen". If the last_price is less than the previous_price, we set the price_change variable to "dropped". If the last_price is equal to the previous_price, we set the price_change variable to "not changed".

Finally, we use an if statement to check if the direction variable is not None. If it is not None, this means that the MACD line has crossed the signal line and we send a message using the telegram.Bot.send_message function. The message includes the direction in which the MACD line crossed the signal line and the change in price since the last message.

After sending the message (if applicable), we use the asyncio.sleep function to pause the program for one minute before repeating the process. This allows us to check for crossings every minute in the 15-minute timeframe.

The main function is then run using the asyncio.run function.

Once the main function has been defined and the while loop has been set up, the program enters an infinite loop. On each iteration of the loop, the program sends a request to the Bybit API to get the MACD and signal line values for the specified timeframe. It then uses the crossover function from the talib library to check if the MACD line has crossed the signal line. If the MACD line has crossed the signal line, the program sets the direction variable to "upwards" or "downwards" depending on the direction of the crossing. If there was no crossing, the direction variable is set to None.

Next, the program checks if the price has risen or dropped since the last message by comparing the last_price variable (which is set to the current close price) to the previous_price variable (which is initially set to None). If the last_price is greater than the previous_price, the price_change variable is set to "risen". If the last_price is less than the previous_price, the price_change variable is set to "dropped". If the last_price is equal to the previous_price, the price_change variable is set to "not changed".

Finally, the program checks if the direction variable is not None. If it is not None, this means that the MACD line has crossed the signal line and the program sends a message using the telegram.Bot.send_message function. The message includes the direction in which the MACD line crossed the signal line and the change in price since the last message.

After sending the message (if applicable), the program uses the asyncio.sleep function to pause the program for one minute before repeating the process. This allows the program to check for crossings every minute in the 15-minute timeframe.