Update byperpv3.py
This code is a script that uses the Python asyncio library to perform a series of tasks asynchronously. The main task is defined in the main() function, which is run using asyncio.run(main()).

The main() function has an infinite loop that runs every minute, using the asyncio.sleep(60) function to pause the loop for one minute.

Inside the loop, the script sends a GET request to the Bybit API to retrieve the latest MACD and signal line values for the specified timeframe. It then uses the talib library's crossover() function to check if the MACD line has crossed the signal line. If the MACD line has crossed the signal line, the script sends a message to a Telegram chat with the direction of the cross and the change in price since the last message.

The telegram library is used to send messages to the Telegram chat using the API token that you provide. The chat_id variable specifies the chat where the messages will be sent.

The previous_price variable is used to track the price change between iterations of the loop. It is initialized to None at the beginning and then updated to the current price each time the loop runs. This allows the script to compare the current price to the previous price to determine if the price has risen or dropped.